+++ src/nebari_plugin_cert_manager_chart/template/chart/templates/certificates.yaml
